+++ src/DetailsView/components/settings-panel/settings/settings-provider-impl.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { BugFilingSettings } from './bug-filing/bug-filing-settings';
+import { BugFilingSettings } from './issue-filing/issue-filing-settings';
 import { HighContrastSettings } from './high-contrast/high-contrast-settings';
 import { createSettingsProvider } from './settings-provider';
 import { TelemetrySettings } from './telemetry/telemetry-settings';
