+++ src/injected/main-window-context.ts
@@ -5,7 +5,7 @@ import { BugActionMessageCreator } from '../common/message-creators/bug-action-m
 import { DevToolActionMessageCreator } from '../common/message-creators/dev-tool-action-message-creator';
 import { DevToolState } from '../common/types/store-data/idev-tool-state';
 import { UserConfigurationStoreData } from '../common/types/store-data/user-configuration-store';
-import { BugFilingServiceProvider } from './../bug-filing/bug-filing-service-provider';
+import { BugFilingServiceProvider } from './../issue-filing/issue-filing-service-provider';
 import { EnvironmentInfoProvider } from './../common/environment-info-provider';
 import { UserConfigMessageCreator } from './../common/message-creators/user-config-message-creator';
 import { TargetPageActionMessageCreator } from './target-page-action-message-creator';
