+++ src/issue-filing/azure-boards/azure-boards-issue-filing-service.tsx
@@ -3,7 +3,7 @@
 import { isEmpty } from 'lodash';
 import { IDropdownOption } from 'office-ui-fabric-react/lib/Dropdown';
 import { createSettingsGetter } from '../common/create-settings-getter';
-import { BugFilingService } from '../types/bug-filing-service';
+import { BugFilingService } from '../types/issue-filing-service';
 import { AzureBoardsSettingsForm } from './azure-boards-settings-form';
 import { azureBoardsIssueFilingUrlProvider } from './create-azure-boards-issue-filing-url';
