+++ src/issue-filing/github/create-github-issue-filing-url.ts
@@ -7,7 +7,7 @@ import { MarkdownFormatter } from '../common/markup/markdown-formatter';
 import { EnvironmentInfo } from './../../common/environment-info-provider';
 import { CreateIssueDetailsTextData } from './../../common/types/create-issue-details-text-data';
 import { IssueFilingUrlStringUtils, IssueUrlCreationUtils } from './../common/issue-filing-url-string-utils';
-import { GitHubBugFilingSettings } from './github-bug-filing-service';
+import { GitHubBugFilingSettings } from './github-issue-filing-service';
 
 export const createGitHubIssueFilingUrlProvider = (
     stringUtils: IssueUrlCreationUtils,
