+++ src/issue-filing/issue-filing-service-provider-impl.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { AzureBoardsBugFilingService } from './azure-boards/azure-boards-bug-filing-service';
-import { BugFilingServiceProvider } from './bug-filing-service-provider';
-import { GitHubBugFilingService } from './github/github-bug-filing-service';
+import { AzureBoardsBugFilingService } from './azure-boards/azure-boards-issue-filing-service';
+import { BugFilingServiceProvider } from './issue-filing-service-provider';
+import { GitHubBugFilingService } from './github/github-issue-filing-service';
 import { NullIssueFilingService } from './null-issue-filing-service/null-issue-filing-service';
 
 export const BugFilingServiceProviderImpl = new BugFilingServiceProvider([
