+++ src/issue-filing/issue-filing-service-provider.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { BugFilingService } from './types/bug-filing-service';
+import { BugFilingService } from './types/issue-filing-service';
 
 export class BugFilingServiceProvider {
     constructor(private readonly services: BugFilingService[]) {}
