+++ src/issue-filing/null-issue-filing-service/null-issue-filing-service.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { NamedSFC } from '../../common/react/named-sfc';
-import { BugFilingService } from '../types/bug-filing-service';
+import { BugFilingService } from '../types/issue-filing-service';
 import { SettingsFormProps } from '../types/settings-form-props';
 
 const nullServiceKey = 'none';
