+++ src/tests/unit/tests/DetailsView/details-view-container.test.tsx
@@ -5,7 +5,7 @@ import { ISelection, Selection } from 'office-ui-fabric-react/lib/DetailsList';
 import * as React from 'react';
 import { IMock, It, Mock, MockBehavior, Times } from 'typemoq';
 
-import { GitHubBugFilingSettings } from '../../../../bug-filing/github/github-bug-filing-service';
+import { GitHubBugFilingSettings } from '../../../../issue-filing/github/github-issue-filing-service';
 import { DropdownClickHandler } from '../../../../common/dropdown-click-handler';
 import { StoreActionMessageCreator } from '../../../../common/message-creators/store-action-message-creator';
 import { StoreActionMessageCreatorImpl } from '../../../../common/message-creators/store-action-message-creator-impl';
