+++ src/tests/unit/tests/bug-filing/azure-boards/create-azure-boards-issue-filing-url.test.ts
@@ -2,12 +2,12 @@
 // Licensed under the MIT License.
 
 import { IMock, Mock } from 'typemoq';
-import { AzureBoardsBugFilingSettings } from '../../../../../bug-filing/azure-boards/azure-boards-bug-filing-service';
-import { createAzureBoardsIssueFilingUrlProvider } from '../../../../../bug-filing/azure-boards/create-azure-boards-issue-filing-url';
-import { HTTPQueryBuilder } from '../../../../../bug-filing/common/http-query-builder';
-import { IssueDetailsBuilder } from '../../../../../bug-filing/common/issue-details-builder';
-import { IssueUrlCreationUtils } from '../../../../../bug-filing/common/issue-filing-url-string-utils';
-import { IssueFilingUrlProvider } from '../../../../../bug-filing/types/bug-filing-service';
+import { AzureBoardsBugFilingSettings } from '../../../../../issue-filing/azure-boards/azure-boards-issue-filing-service';
+import { createAzureBoardsIssueFilingUrlProvider } from '../../../../../issue-filing/azure-boards/create-azure-boards-issue-filing-url';
+import { HTTPQueryBuilder } from '../../../../../issue-filing/common/http-query-builder';
+import { IssueDetailsBuilder } from '../../../../../issue-filing/common/issue-details-builder';
+import { IssueUrlCreationUtils } from '../../../../../issue-filing/common/issue-filing-url-string-utils';
+import { IssueFilingUrlProvider } from '../../../../../issue-filing/types/issue-filing-service';
 import { EnvironmentInfo } from '../../../../../common/environment-info-provider';
 import { title } from '../../../../../content/strings/application';
