+++ src/tests/unit/tests/bug-filing/common/issue-details-builder.test.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { IMock, It, Mock, MockBehavior } from 'typemoq';
-import { createIssueDetailsBuilder } from '../../../../../bug-filing/common/create-issue-details-builder';
-import { MarkupFormatter } from '../../../../../bug-filing/common/markup/markup-formatter';
+import { createIssueDetailsBuilder } from '../../../../../issue-filing/common/create-issue-details-builder';
+import { MarkupFormatter } from '../../../../../issue-filing/common/markup/markup-formatter';
 import { EnvironmentInfo } from '../../../../../common/environment-info-provider';
 
 describe('Name of the group', () => {
