+++ src/tests/unit/tests/bug-filing/common/markup/truncate-snippet.test.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import { repeat } from 'lodash';
 
-import { maxSnippetLength, truncateSnippet } from '../../../../../../bug-filing/common/markup/truncate-snippet';
+import { maxSnippetLength, truncateSnippet } from '../../../../../../issue-filing/common/markup/truncate-snippet';
 
 describe('truncateSnippet', () => {
     const testSubject = truncateSnippet;
