+++ src/tests/unit/tests/bug-filing/github/github-issue-filing-service.test.tsx
@@ -5,10 +5,10 @@ import { TextField } from 'office-ui-fabric-react/lib/TextField';
 import * as React from 'react';
 import { IMock, Mock, Times } from 'typemoq';
 
-import { OnPropertyUpdateCallback } from '../../../../../bug-filing/components/bug-filing-settings-container';
-import { gitHubIssueFilingUrlProvider } from '../../../../../bug-filing/github/create-github-bug-filing-url';
-import { GitHubBugFilingService, GitHubBugFilingSettings } from '../../../../../bug-filing/github/github-bug-filing-service';
-import { SettingsFormProps } from '../../../../../bug-filing/types/settings-form-props';
+import { OnPropertyUpdateCallback } from '../../../../../issue-filing/components/issue-filing-settings-container';
+import { gitHubIssueFilingUrlProvider } from '../../../../../issue-filing/github/create-github-issue-filing-url';
+import { GitHubBugFilingService, GitHubBugFilingSettings } from '../../../../../issue-filing/github/github-issue-filing-service';
+import { SettingsFormProps } from '../../../../../issue-filing/types/settings-form-props';
 import { BugServicePropertiesMap } from '../../../../../common/types/store-data/user-configuration-store';
 import { SettingsDeps } from '../../../../../DetailsView/components/settings-panel/settings/settings-props';
